sprint-6-pb-aws-furg-ifrs-uffs/equipe-2/api-tts/requirements.txt
python-dotenv==1.0.0
boto3==1.28.39